sql: split types.go into smaller files
Types.go was starting to get unmanagable so I've taken any non-standard type
and pulled it into its own file.

Release note: None
