+++ pkg/sql/sem/types/types.go
@@ -15,7 +15,6 @@
 package types
 
 import (
-	"bytes"
 	"fmt"
 
 	"github.com/lib/pq/oid"
@@ -185,45 +184,6 @@ func (tString) Oid() oid.Oid             { return oid.T_text }
 func (tString) SQLName() string          { return "text" }
 func (tString) IsAmbiguous() bool        { return false }
 
-// TCollatedString is the type of strings with a locale.
-type TCollatedString struct {
-	Locale string
-}
-
-// String implements the fmt.Stringer interface.
-func (t TCollatedString) String() string {
-	return fmt.Sprintf("collatedstring{%s}", t.Locale)
-}
-
-// Equivalent implements the T interface.
-func (t TCollatedString) Equivalent(other T) bool {
-	if other == Any {
-		return true
-	}
-	u, ok := UnwrapType(other).(TCollatedString)
-	if ok {
-		return t.Locale == "" || u.Locale == "" || t.Locale == u.Locale
-	}
-	return false
-}
-
-// FamilyEqual implements the T interface.
-func (TCollatedString) FamilyEqual(other T) bool {
-	_, ok := UnwrapType(other).(TCollatedString)
-	return ok
-}
-
-// Oid implements the T interface.
-func (TCollatedString) Oid() oid.Oid { return oid.T_text }
-
-// SQLName implements the T interface.
-func (TCollatedString) SQLName() string { return "text" }
-
-// IsAmbiguous implements the T interface.
-func (t TCollatedString) IsAmbiguous() bool {
-	return t.Locale == ""
-}
-
 type tBytes struct{}
 
 func (tBytes) String() string           { return "bytes" }
@@ -316,167 +276,6 @@ func (tINet) FamilyEqual(other T) bool { return UnwrapType(other) == INet }
 func (tINet) Oid() oid.Oid             { return oid.T_inet }
 func (tINet) SQLName() string          { return "inet" }
 func (tINet) IsAmbiguous() bool        { return false }
-
-// TTuple is the type of a DTuple.
-type TTuple struct {
-	Types  []T
-	Labels []string
-}
-
-// String implements the fmt.Stringer interface.
-func (t TTuple) String() string {
-	var buf bytes.Buffer
-	buf.WriteString("tuple")
-	if t.Types != nil {
-		buf.WriteByte('{')
-		for i, typ := range t.Types {
-			if i != 0 {
-				buf.WriteString(", ")
-			}
-			buf.WriteString(typ.String())
-			if t.Labels != nil {
-				buf.WriteString(" AS ")
-				buf.WriteString(t.Labels[i])
-			}
-		}
-		buf.WriteByte('}')
-	}
-	return buf.String()
-}
-
-// Equivalent implements the T interface.
-func (t TTuple) Equivalent(other T) bool {
-	if other == Any {
-		return true
-	}
-	u, ok := UnwrapType(other).(TTuple)
-	if !ok {
-		return false
-	}
-	if len(t.Types) == 0 || len(u.Types) == 0 {
-		// Tuples that aren't fully specified (have a nil subtype list) are always
-		// equivalent to other tuples, to allow overloads to specify that they take
-		// an arbitrary tuple type.
-		return true
-	}
-	if len(t.Types) != len(u.Types) {
-		return false
-	}
-	for i, typ := range t.Types {
-		if !typ.Equivalent(u.Types[i]) {
-			return false
-		}
-	}
-	return true
-}
-
-// FamilyEqual implements the T interface.
-func (TTuple) FamilyEqual(other T) bool {
-	_, ok := UnwrapType(other).(TTuple)
-	return ok
-}
-
-// Oid implements the T interface.
-func (TTuple) Oid() oid.Oid { return oid.T_record }
-
-// SQLName implements the T interface.
-func (TTuple) SQLName() string { return "record" }
-
-// IsAmbiguous implements the T interface.
-func (t TTuple) IsAmbiguous() bool {
-	for _, typ := range t.Types {
-		if typ == nil || typ.IsAmbiguous() {
-			return true
-		}
-	}
-	return len(t.Types) == 0
-}
-
-// TPlaceholder is the type of a placeholder.
-type TPlaceholder struct {
-	Name string
-}
-
-// String implements the fmt.Stringer interface.
-func (t TPlaceholder) String() string { return fmt.Sprintf("placeholder{%s}", t.Name) }
-
-// Equivalent implements the T interface.
-func (t TPlaceholder) Equivalent(other T) bool {
-	if other == Any {
-		return true
-	}
-	if other.IsAmbiguous() {
-		return true
-	}
-	u, ok := UnwrapType(other).(TPlaceholder)
-	return ok && t.Name == u.Name
-}
-
-// FamilyEqual implements the T interface.
-func (TPlaceholder) FamilyEqual(other T) bool {
-	_, ok := UnwrapType(other).(TPlaceholder)
-	return ok
-}
-
-// Oid implements the T interface.
-func (TPlaceholder) Oid() oid.Oid { panic("TPlaceholder.Oid() is undefined") }
-
-// SQLName implements the T interface.
-func (TPlaceholder) SQLName() string { panic("TPlaceholder.SQLName() is undefined") }
-
-// IsAmbiguous implements the T interface.
-func (TPlaceholder) IsAmbiguous() bool { panic("TPlaceholder.IsAmbiguous() is undefined") }
-
-// TArray is the type of a DArray.
-type TArray struct{ Typ T }
-
-func (a TArray) String() string { return a.Typ.String() + "[]" }
-
-// Equivalent implements the T interface.
-func (a TArray) Equivalent(other T) bool {
-	if other == Any {
-		return true
-	}
-	if u, ok := UnwrapType(other).(TArray); ok {
-		return a.Typ.Equivalent(u.Typ)
-	}
-	return false
-}
-
-// FamilyEqual implements the T interface.
-func (TArray) FamilyEqual(other T) bool {
-	_, ok := UnwrapType(other).(TArray)
-	return ok
-}
-
-const noArrayType = 0
-
-// ArrayOids is a set of all oids which correspond to an array type.
-var ArrayOids = map[oid.Oid]struct{}{}
-
-func init() {
-	for _, v := range oidToArrayOid {
-		ArrayOids[v] = struct{}{}
-	}
-}
-
-// Oid implements the T interface.
-func (a TArray) Oid() oid.Oid {
-	if o, ok := oidToArrayOid[a.Typ.Oid()]; ok {
-		return o
-	}
-	return noArrayType
-}
-
-// SQLName implements the T interface.
-func (a TArray) SQLName() string {
-	return a.Typ.SQLName() + "[]"
-}
-
-// IsAmbiguous implements the T interface.
-func (a TArray) IsAmbiguous() bool {
-	return a.Typ == nil || a.Typ.IsAmbiguous()
-}
 
 type tAny struct{}
